+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog for PInvokeHelper
 
+## v0.2.0 - TBD
 
-## v0.0.1 - 2021-05-01
+* Changed `Open-ProcessToken` to `Get-ProcessToken`
+* Added the following cmdlet:
+  * [Get-CurrentThreadId](docs/en-US/Get-CurrentThreadId.md) - Gets the Thread ID of the current thread
+  * [Get-ThreadToken](docs/en-US/Get-ThreadToken.md) - Gets the access token for the thread specified
+
+## v0.1.0 - 2021-05-01
 
 * Initial version of the `PSAccessToken` module